% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_crosswalk.R
\docType{data}
\name{BETA_TABLES}
\alias{BETA_TABLES}
\title{The 16 CLIF beta tables}
\format{
An object of class \code{character} of length 16.
}
\usage{
BETA_TABLES
}
\description{
The default scope for a 2.1 to 3.0 migration.
}
\keyword{datasets}
