% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_validator.R
\name{check_categorical_values}
\alias{check_categorical_values}
\title{C.5 Check categorical values against mCIDE permissible values}
\usage{
check_categorical_values(df, schema, table_name)
}
\arguments{
\item{df}{Data to validate.}

\item{schema}{Parsed table schema containing \code{required_columns}.}

\item{table_name}{Name of the table.}
}
\value{
A \code{dqa_conformance_result}.
}
\description{
C.5 Check categorical values against mCIDE permissible values
}
\examples{
\dontrun{
check_categorical_values(patient_data, load_schema("patient"), "patient")
}
}
