% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_validator.R
\name{check_cross_table_temporal_plausibility}
\alias{check_cross_table_temporal_plausibility}
\title{P.8 Check that event times fall within hospitalization bounds}
\usage{
check_cross_table_temporal_plausibility(
  target_df,
  hospitalization_df,
  target_table,
  time_columns,
  warning_threshold = 0,
  error_threshold = 10
)
}
\arguments{
\item{target_df}{The table whose timestamps are being checked.}

\item{hospitalization_df}{The hospitalization table supplying admission and
discharge bounds.}

\item{target_table}{Name of the target table.}

\item{time_columns}{Character vector of datetime columns to check.}

\item{warning_threshold}{Violation percent above which a warning is raised.}

\item{error_threshold}{Violation percent above which an error is raised.}
}
\value{
A \code{dqa_plausibility_result}.
}
\description{
P.8 Check that event times fall within hospitalization bounds
}
\examples{
\dontrun{
check_cross_table_temporal_plausibility(labs_data, hosp_data, "labs",
                                        c("lab_collect_dttm"))
}
}
