% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_validator.R
\name{check_table_presence}
\alias{check_table_presence}
\title{C.1b Check that a loaded table has rows and columns}
\usage{
check_table_presence(df, table_name)
}
\arguments{
\item{df}{Data to validate.}

\item{table_name}{Name of the table.}
}
\value{
A \code{dqa_conformance_result}.
}
\description{
C.1b Check that a loaded table has rows and columns
}
\examples{
check_table_presence(data.frame(a = 1), "patient")
}
