% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_stitching_encounters.R
\name{clif-stitching-encounters}
\alias{clif-stitching-encounters}
\title{Encounter stitching utilities}
\description{
Port of \code{clifpy/utils/stitching_encounters.py} (clifpy 0.5.0). Groups
hospitalizations that occur within a configurable number of hours of each
other into a single \code{encounter_block}, so that rapid discharge/readmission
sequences (e.g. ED to inpatient transfers) can be analyzed as one continuous
encounter.
}
