% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_registry.R
\docType{data}
\name{CLIF_TABLE_NAMES}
\alias{CLIF_TABLE_NAMES}
\title{Names of the CLIF tables clifR implements}
\format{
An object of class \code{character} of length 18.
}
\usage{
CLIF_TABLE_NAMES
}
\description{
Ordered to match clifpy's \code{TABLE_CLASSES} so that any iteration over tables
produces the same sequence in both languages.
}
\keyword{datasets}
