% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_code_status.R
\name{CodeStatus}
\alias{CodeStatus}
\title{CodeStatus table}
\description{
R6 class for the CLIF code_status table (resuscitation preferences such as DNR, DNR/DNI and Full Code).
Inherits all loading, validation and summary behaviour from \link{BaseTable}.
Port of \code{clifpy.tables.code_status.CodeStatus}.
}
\examples{
\dontrun{
code_status_table <- CodeStatus$new(
  data_directory = "data/clif",
  filetype = "parquet",
  timezone = "US/Central"
)
code_status_table$validate()
}
}
\section{Super class}{
\code{\link[clifR:BaseTable]{clifR::BaseTable}} -> \code{CodeStatus}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CodeStatus-new}{\code{CodeStatus$new()}}
\item \href{#method-CodeStatus-clone}{\code{CodeStatus$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="clifR" data-topic="BaseTable" data-id="analyze_categorical_distributions"><a href='../../clifR/html/BaseTable.html#method-BaseTable-analyze_categorical_distributions'><code>clifR::BaseTable$analyze_categorical_distributions()</code></a></span></li>
<li><span class="pkg-link" data-pkg="clifR" data-topic="BaseTable" data-id="calculate_stratified_ecdf"><a href='../../clifR/html/BaseTable.html#method-BaseTable-calculate_stratified_ecdf'><code>clifR::BaseTable$calculate_stratified_ecdf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="clifR" data-topic="BaseTable" data-id="get_summary"><a href='../../clifR/html/BaseTable.html#method-BaseTable-get_summary'><code>clifR::BaseTable$get_summary()</code></a></span></li>
<li><span class="pkg-link" data-pkg="clifR" data-topic="BaseTable" data-id="isvalid"><a href='../../clifR/html/BaseTable.html#method-BaseTable-isvalid'><code>clifR::BaseTable$isvalid()</code></a></span></li>
<li><span class="pkg-link" data-pkg="clifR" data-topic="BaseTable" data-id="plot_categorical_distributions"><a href='../../clifR/html/BaseTable.html#method-BaseTable-plot_categorical_distributions'><code>clifR::BaseTable$plot_categorical_distributions()</code></a></span></li>
<li><span class="pkg-link" data-pkg="clifR" data-topic="BaseTable" data-id="print"><a href='../../clifR/html/BaseTable.html#method-BaseTable-print'><code>clifR::BaseTable$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="clifR" data-topic="BaseTable" data-id="save_summary"><a href='../../clifR/html/BaseTable.html#method-BaseTable-save_summary'><code>clifR::BaseTable$save_summary()</code></a></span></li>
<li><span class="pkg-link" data-pkg="clifR" data-topic="BaseTable" data-id="validate"><a href='../../clifR/html/BaseTable.html#method-BaseTable-validate'><code>clifR::BaseTable$validate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CodeStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-CodeStatus-new}{}}}
\subsection{Method \code{new()}}{
Create a CodeStatus table instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CodeStatus$new(
  data_directory = NULL,
  filetype = NULL,
  timezone = "UTC",
  output_directory = NULL,
  data = NULL,
  clif_version = DEFAULT_CLIF_VERSION
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_directory}}{Directory containing the CLIF data files.}

\item{\code{filetype}}{Either \code{"csv"} or \code{"parquet"}.}

\item{\code{timezone}}{Olson timezone for datetime columns.}

\item{\code{output_directory}}{Directory for logs and outputs. Defaults to
\verb{<data_directory>/output}.}

\item{\code{data}}{Optional pre-loaded data frame. When supplied, no file is read.}

\item{\code{clif_version}}{CLIF schema version. Defaults to \link{DEFAULT_CLIF_VERSION}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{CodeStatus} instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CodeStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CodeStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CodeStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
