% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_crosswalk.R
\name{crosswalk_target_columns}
\alias{crosswalk_target_columns}
\title{Standardized columns present in a set of available columns}
\usage{
crosswalk_target_columns(table_name, available_columns)
}
\arguments{
\item{table_name}{snake_case CLIF table name.}

\item{available_columns}{Character vector of column names present in the data.}
}
\value{
Character vector, in \code{available_columns} order.
}
\description{
Standardized columns present in a set of available columns
}
\keyword{internal}
