% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas.R
\docType{data}
\name{DEFAULT_CLIF_VERSION}
\alias{DEFAULT_CLIF_VERSION}
\title{Default CLIF version}
\format{
An object of class \code{character} of length 1.
}
\usage{
DEFAULT_CLIF_VERSION
}
\description{
Used when a caller does not specify one. Kept at 2.1 to match clifpy's default.
}
\keyword{datasets}
