% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_data.R
\docType{data}
\name{DEMO_TABLES}
\alias{DEMO_TABLES}
\title{Tables available in the bundled demo dataset}
\format{
An object of class \code{character} of length 9.
}
\usage{
DEMO_TABLES
}
\description{
Tables available in the bundled demo dataset
}
\keyword{datasets}
