% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_sofa.R
\docType{data}
\name{DEVICE_RANK_DICT}
\alias{DEVICE_RANK_DICT}
\title{Respiratory device severity ranks (lower rank = more support)}
\format{
An object of class \code{integer} of length 9.
}
\usage{
DEVICE_RANK_DICT
}
\description{
Respiratory device severity ranks (lower rank = more support)
}
\keyword{datasets}
