% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_sofa.R
\docType{data}
\name{DEVICE_RANK_MAPPING}
\alias{DEVICE_RANK_MAPPING}
\title{Device rank lookup table}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9 rows and 2 columns.
}
\usage{
DEVICE_RANK_MAPPING
}
\description{
Tibble form of \link{DEVICE_RANK_DICT}, joined onto the wide dataset so that the
worst (lowest-ranked) device per ID can be recovered after aggregation.
}
\keyword{datasets}
