% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_io.R
\name{export_to_json}
\alias{export_to_json}
\title{Export data to JSON}
\usage{
export_to_json(data, file_path, pretty = TRUE)
}
\arguments{
\item{data}{tibble or data.frame. Data to export.}

\item{file_path}{Character. Output JSON file path.}

\item{pretty}{Logical. Pretty print JSON (default: TRUE).}
}
\value{
Invisible NULL.
}
\description{
Export data to JSON
}
