% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_wide_dataset.R
\name{get_timestamp_column}
\alias{get_timestamp_column}
\title{Timestamp column configured for a table}
\usage{
get_timestamp_column(table_name)
}
\arguments{
\item{table_name}{snake_case table name.}
}
\value{
Column name, or \code{NULL}.
}
\description{
Timestamp column configured for a table
}
\keyword{internal}
