% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_data.R
\name{load_demo_clif}
\alias{load_demo_clif}
\title{Load an orchestrator populated with demo data}
\usage{
load_demo_clif(tables = NULL, timezone = "UTC", verbose = FALSE)
}
\arguments{
\item{tables}{Character vector of tables to load. Defaults to all of \link{DEMO_TABLES}.}

\item{timezone}{Olson timezone for datetime columns.}

\item{verbose}{Whether to report which tables were loaded.}
}
\value{
A \link{ClifOrchestrator} with the requested demo tables loaded.
}
\description{
Load an orchestrator populated with demo data
}
\examples{
demo_orchestrator <- load_demo_clif(tables = c("patient", "vitals"))
demo_orchestrator$get_loaded_tables()
}
