% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_logging_config.R
\name{log_validation_header}
\alias{log_validation_header}
\title{Print a validation report header}
\usage{
log_validation_header(table_name, n_rows)
}
\arguments{
\item{table_name}{Name of the table being validated.}

\item{n_rows}{Number of rows in the table.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Print a validation report header
}
\keyword{internal}
