% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_sofa.R
\docType{data}
\name{MAX_ITEMS}
\alias{MAX_ITEMS}
\title{Wide-dataset columns aggregated with MAX when taking the worst value}
\format{
An object of class \code{character} of length 7.
}
\usage{
MAX_ITEMS
}
\description{
Wide-dataset columns aggregated with MAX when taking the worst value
}
\keyword{datasets}
