% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_sofa.R
\docType{data}
\name{MIN_ITEMS}
\alias{MIN_ITEMS}
\title{Wide-dataset columns aggregated with MIN when taking the worst value}
\format{
An object of class \code{character} of length 6.
}
\usage{
MIN_ITEMS
}
\description{
Wide-dataset columns aggregated with MIN when taking the worst value
}
\keyword{datasets}
