% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_table.R
\name{null-coalesce}
\alias{null-coalesce}
\alias{\%||\%}
\title{Null-coalescing operator}
\usage{
lhs \%||\% rhs
}
\arguments{
\item{lhs}{Value to use when not \code{NULL}.}

\item{rhs}{Fallback value.}
}
\value{
\code{lhs} when it is not \code{NULL}, otherwise \code{rhs}.
}
\description{
Null-coalescing operator
}
\keyword{internal}
