% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_query.R
\name{query_dict_sql_clause}
\alias{query_dict_sql_clause}
\title{Build the SQL predicate for a whole query specification}
\usage{
query_dict_sql_clause(query_dict, category_column)
}
\arguments{
\item{query_dict}{Named list mapping category values to extremal-value vectors.}

\item{category_column}{Name of the table's category column.}
}
\value{
A SQL boolean expression.
}
\description{
Build the SQL predicate for a whole query specification
}
\keyword{internal}
