% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_query.R
\docType{data}
\name{QUERY_SUPPORTED_TABLE_NAMES}
\alias{QUERY_SUPPORTED_TABLE_NAMES}
\title{Table names supported by \code{\link[=lookup_extremal_values_in_long_table]{lookup_extremal_values_in_long_table()}}}
\format{
An object of class \code{character} of length 4.
}
\usage{
QUERY_SUPPORTED_TABLE_NAMES
}
\description{
Table names supported by \code{\link[=lookup_extremal_values_in_long_table]{lookup_extremal_values_in_long_table()}}
}
\keyword{datasets}
