% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_query.R
\docType{data}
\name{QUERY_TABLE_DTTM_COLUMNS}
\alias{QUERY_TABLE_DTTM_COLUMNS}
\title{Long-table datetime column by table name}
\format{
An object of class \code{character} of length 4.
}
\usage{
QUERY_TABLE_DTTM_COLUMNS
}
\description{
Long-table datetime column by table name
}
\keyword{internal}
