% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_sofa_direct.R
\docType{data}
\name{REQUIRED_ASSESSMENTS}
\alias{REQUIRED_ASSESSMENTS}
\title{Assessment categories required by \code{\link[=compute_sofa_direct]{compute_sofa_direct()}}}
\format{
An object of class \code{character} of length 1.
}
\usage{
REQUIRED_ASSESSMENTS
}
\description{
Assessment categories required by \code{\link[=compute_sofa_direct]{compute_sofa_direct()}}
}
\keyword{internal}
