% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_sofa_direct.R
\docType{data}
\name{REQUIRED_LABS}
\alias{REQUIRED_LABS}
\title{Lab categories required by \code{\link[=compute_sofa_direct]{compute_sofa_direct()}}}
\format{
An object of class \code{character} of length 4.
}
\usage{
REQUIRED_LABS
}
\description{
Lab categories required by \code{\link[=compute_sofa_direct]{compute_sofa_direct()}}
}
\keyword{internal}
