% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_sofa_direct.R
\docType{data}
\name{REQUIRED_MEDS}
\alias{REQUIRED_MEDS}
\title{Medication categories required by \code{\link[=compute_sofa_direct]{compute_sofa_direct()}}}
\format{
An object of class \code{character} of length 4.
}
\usage{
REQUIRED_MEDS
}
\description{
Medication categories required by \code{\link[=compute_sofa_direct]{compute_sofa_direct()}}
}
\keyword{internal}
