% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_sofa_direct.R
\docType{data}
\name{REQUIRED_RESP_SUPPORT_COLS}
\alias{REQUIRED_RESP_SUPPORT_COLS}
\title{Respiratory support columns required by \code{\link[=compute_sofa_direct]{compute_sofa_direct()}}}
\format{
An object of class \code{character} of length 3.
}
\usage{
REQUIRED_RESP_SUPPORT_COLS
}
\description{
Respiratory support columns required by \code{\link[=compute_sofa_direct]{compute_sofa_direct()}}
}
\keyword{internal}
