% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_sofa.R
\docType{data}
\name{REQUIRED_SOFA_CATEGORIES_BY_TABLE}
\alias{REQUIRED_SOFA_CATEGORIES_BY_TABLE}
\title{Categories required from each CLIF table for SOFA scoring}
\format{
An object of class \code{list} of length 5.
}
\usage{
REQUIRED_SOFA_CATEGORIES_BY_TABLE
}
\description{
Named list mirroring clifpy's \code{REQUIRED_SOFA_CATEGORIES_BY_TABLE}; used as the
\code{category_filters} argument when building a wide dataset for SOFA.
}
\keyword{datasets}
