% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_sofa_direct.R
\docType{data}
\name{REQUIRED_VITALS}
\alias{REQUIRED_VITALS}
\title{Vital categories required by \code{\link[=compute_sofa_direct]{compute_sofa_direct()}}}
\format{
An object of class \code{character} of length 3.
}
\usage{
REQUIRED_VITALS
}
\description{
Vital categories required by \code{\link[=compute_sofa_direct]{compute_sofa_direct()}}
}
\keyword{internal}
