% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_wide_dataset.R
\name{save_wide_dataset}
\alias{save_wide_dataset}
\title{Write a wide dataset to the data directory}
\usage{
save_wide_dataset(data, data_directory, output_filename, output_format)
}
\arguments{
\item{data}{Wide dataset to write.}

\item{data_directory}{Destination directory.}

\item{output_filename}{File name stem; defaults to \code{wide_dataset_YYYYMMDD_HHMMSS}.}

\item{output_format}{\code{"csv"} or \code{"parquet"}.}
}
\value{
Path to the written file, invisibly.
}
\description{
Write a wide dataset to the data directory
}
\keyword{internal}
