% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_table.R
\name{schema_category_columns}
\alias{schema_category_columns}
\title{Column names flagged as categorical in a schema}
\usage{
schema_category_columns(schema)
}
\arguments{
\item{schema}{Parsed table schema.}
}
\value{
Character vector of category column names.
}
\description{
Column names flagged as categorical in a schema
}
\keyword{internal}
