% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas.R
\name{schema_dir}
\alias{schema_dir}
\title{Path to the schema directory for a CLIF version}
\usage{
schema_dir(clif_version = DEFAULT_CLIF_VERSION)
}
\arguments{
\item{clif_version}{CLIF version string. Defaults to \link{DEFAULT_CLIF_VERSION}.}
}
\value{
Absolute path to the version's schema directory.
}
\description{
Path to the schema directory for a CLIF version
}
