% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_registry.R
\name{snake_to_pascal_case}
\alias{snake_to_pascal_case}
\title{Convert a snake_case name to PascalCase}
\usage{
snake_to_pascal_case(snake_case_name)
}
\arguments{
\item{snake_case_name}{e.g. \code{"respiratory_support"}.}
}
\value{
e.g. \code{"RespiratorySupport"}.
}
\description{
Convert a snake_case name to PascalCase
}
\keyword{internal}
