% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_sofa.R
\docType{data}
\name{SOFA_SUBSCORE_COLUMNS}
\alias{SOFA_SUBSCORE_COLUMNS}
\title{Names of the six SOFA component score columns}
\format{
An object of class \code{character} of length 6.
}
\usage{
SOFA_SUBSCORE_COLUMNS
}
\description{
Names of the six SOFA component score columns
}
\keyword{internal}
