% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_io.R
\name{sql_quote_value}
\alias{sql_quote_value}
\title{Quote a value for inline SQL}
\usage{
sql_quote_value(value)
}
\arguments{
\item{value}{A scalar to embed in a SQL string literal.}
}
\value{
The value wrapped in single quotes with internal quotes escaped.
}
\description{
Quote a value for inline SQL
}
\keyword{internal}
