% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas.R
\docType{data}
\name{SUPPORTED_CLIF_VERSIONS}
\alias{SUPPORTED_CLIF_VERSIONS}
\title{CLIF versions for which a schema subdirectory exists}
\format{
An object of class \code{character} of length 2.
}
\usage{
SUPPORTED_CLIF_VERSIONS
}
\description{
CLIF versions for which a schema subdirectory exists
}
\keyword{datasets}
