% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clifR-package.R
\docType{package}
\name{clifR-package}
\alias{clifR}
\alias{clifR-package}
\title{clifR: Common Longitudinal ICU data Format Tools for R}
\description{
An R port of the Python library clifpy for standardizing and analyzing critical care (ICU) data using the Common Longitudinal ICU data Format (CLIF) version 2.0. Provides tools for data validation, unit conversion, clinical score calculation (SOFA, Charlson Comorbidity Index), wide dataset transformation, and encounter stitching. Designed for high-performance processing of large ICU datasets with built-in schema validation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AartikSarma/clifR}
  \item Report bugs at \url{https://github.com/AartikSarma/clifR/issues}
}

}
\author{
\strong{Maintainer}: Aartik Sarma \email{aartiksarma@example.com}

Authors:
\itemize{
  \item CLIF Consortium
}

}
\keyword{internal}
